+++ wattsup/energymon-wattsup.c
@@ -32,12 +32,18 @@ int energymon_get_default(energymon* em) {
 // Keeping this as an undocumented feature for now.
 #define ENERGYMON_WATTSUP_ENABLE_ESTIMATES "ENERGYMON_WATTSUP_ENABLE_ESTIMATES"
 
-// must wait at least one second between WattsUp polls
+// WattsUp values refresh every second
 #define WU_MIN_INTERVAL_US 1000000
+// we will poll the device 10x faster - data is often available (even if it hasn't changed)
+#define WU_POLL_INTERVAL_US 100000
 #define WU_POWER_INDEX 3
+// buffer is large enough to hold a handful of complete data packets (usually ~80 bytes)
 #define WU_BUFSIZE 256
-// chosen empirically as a ridiculous value - never needed more than ~4 reads
-#define WU_MAX_RETRIES 5
+// max number of attempts to find a complete data packet during init
+#define WU_INIT_MAX_RETRIES 5
+// configurations for handling incomplete data packets
+#define WU_PACKET_MAX_RETRIES 10
+#define WU_PACKET_WAIT_INTERVAL_US 10000
 
 typedef struct energymon_wattsup {
   energymon_wattsup_ctx* ctx;
@@ -72,79 +78,141 @@ static void lock_release(int* lock) {
 #endif
 }
 
+// Only for use by the polling thread - enables pthread cancel while sleeping, then disables it
+static int wattsup_thread_sleep_us(uint64_t us, volatile const int* poll) {
+  assert(poll != NULL);
+  int ret = 0;
+  if (*poll) {
+#ifndef __ANDROID__
+    // Deadlock can occur during disconnect in some wattsup_driver impls if thread is canceled during I/O
+    // Enable thread cancel while sleeping, disable during I/O and when holding locks
+    int dummy_old_state;
+    pthread_setcancelstate(PTHREAD_CANCEL_ENABLE, &dummy_old_state);
+#endif
+    ret = energymon_sleep_us(us, poll);
+#ifndef __ANDROID__
+    pthread_setcancelstate(PTHREAD_CANCEL_DISABLE, &dummy_old_state);
+#endif
+  }
+  return ret;
+}
+
+// buf must be the start of a valid packet, i.e., begin with a '#'
+static int data_packet_parse(char* buf, unsigned int* deciwatts) {
+  assert(buf != NULL);
+  assert(buf[0] == '#');
+  assert(deciwatts != NULL);
+  unsigned int tmp;
+  int i;
+  char* saveptr;
+  char* tok = strtok_r(buf, ",", &saveptr);
+  for (i = 0; tok; i++) {
+    // 1st index = command, starting with a '#'
+    // 2nd index = subcommand (should be a '-', but we don't care)
+    if (i == WU_POWER_INDEX) {
+      // treat this is as the average power since last successful read
+      errno = 0;
+      tmp = strtoul(tok, NULL, 0);
+      if (errno) {
+        // keep old value of deciwatts
+        perror("data_packet_parse: strtoul");
+        return -1;
+      }
+      *deciwatts = tmp;
+      // ignore the rest of the packet
+      return 0;
+    }
+    tok = strtok_r(NULL, ",", &saveptr);
+  }
+  errno = EBADMSG;
+  perror("data_packet_parse: Syntax error while parsing WattsUp data packet");
+  return -1;
+}
+
+// Returns a pointer (in buf) to the start of a valid packet, NULL otherwise
+static char* data_packet_read(energymon_wattsup_ctx* ctx, char* buf, size_t bufsize, const int* poll) {
+  assert(ctx != NULL);
+  assert(buf != NULL);
+  assert(poll != NULL);
+  char* start;
+  size_t len;
+  int ret;
+  int i = 0;
+  if ((ret = wattsup_read(ctx, buf, bufsize - 1)) < 0) {
+    perror("data_packet_read: wattsup_read");
+    return NULL;
+  }
+  buf[ret] = '\0';
+#ifdef ENERGYMON_WATTSUP_DEBUG
+  fprintf(stdout, "Read %d characters:\n%s\n", ret, buf);
+#endif
+  if (!(*poll)) {
+    // we were probably ordered to stop during I/O
+    return NULL;
+  }
+  // try to find the start of the latest packet (ignore any prior packets that might be in the buffer)
+  if (!(start = strrchr(buf, '#'))) {
+    // first 2 bytes may be modem status codes (libftdi handles this for us)
+    if (ret > 2) {
+      // we probably got the remainder of a previously incomplete packet
+      errno = EBADMSG;
+      perror("Invalid data packet from WattsUp");
+    } else {
+      // data is not ready - this is common when polling faster than device's refresh rate
+      errno = ENODATA;
+    }
+    // nothing we can do except try again later once the device refreshes
+    return NULL;
+  }
+  // check for the end of the data packet (';')
+  for (i = 0; !strchr(start, ';') && i < WU_PACKET_MAX_RETRIES; i++) {
+    if ((len = strlen(buf)) >= bufsize - 1) {
+      errno = ENOBUFS;
+      perror("data_packet_read: Too much data");
+      return NULL;
+    }
+    // short wait before reading again to try and get the rest of the packet
+    wattsup_thread_sleep_us(WU_PACKET_WAIT_INTERVAL_US, poll);
+    if (!(*poll)) {
+      // we were probably ordered to stop during I/O or sleep
+      return NULL;
+    }
+    if ((ret = wattsup_read(ctx, buf + len, bufsize - len - 1)) < 0) {
+      perror("data_packet_read: wattsup_read");
+      return NULL;
+    }
+    buf[len + ret] = '\0';
+#ifdef ENERGYMON_WATTSUP_DEBUG
+    fprintf(stdout, "Read %d additional characters:\n%s\n", ret, buf + len);
+#endif
+  }
+  if (i >= WU_PACKET_MAX_RETRIES) {
+    fprintf(stderr, "Failed to get complete data packet from WattsUp\n");
+    errno = ENODATA;
+    return NULL;
+  }
+  return start;
+}
+
 /**
  * pthread function to poll the device at regular intervals.
+ * Always makes a best-effort to not read or sleep once ordered to stop (state->poll = 0).
  */
 static void* wattsup_poll_sensors(void* args) {
   energymon_wattsup* state = (energymon_wattsup*) args;
-  unsigned int i;
-  unsigned int tmp;
   char buf[WU_BUFSIZE] = { 0 };
-  char* saveptr;
-  char* tok;
-  int ret;
-#ifndef __ANDROID__
-  int dummy_old_state; // not used but kept for portability (see pthread docs)
-#endif
-
+  char* pstart;
   state->deciwatts = 0;
   if (!(state->last_us = energymon_gettime_us())) {
     // must be that CLOCK_MONOTONIC is not supported
     perror("wattsup_poll_sensors");
     return (void*) NULL;
   }
-  energymon_sleep_us(WU_MIN_INTERVAL_US, &state->poll);
+  wattsup_thread_sleep_us(WU_POLL_INTERVAL_US, &state->poll);
   while (state->poll) {
-#ifndef __ANDROID__
-    // deadlock can occur during disconnect in some wattsup_driver impls if thread is canceled during I/O
-    // disable thread cancel while we do I/O and later while we hold the lock
-    pthread_setcancelstate(PTHREAD_CANCEL_DISABLE, &dummy_old_state);
-#endif
-    ret = wattsup_read(state->ctx, buf, WU_BUFSIZE);
-    if (ret < 0) {
-      perror("wattsup_poll_sensors:wattsup_read");
-    } else {
-      if (ret < WU_BUFSIZE) {
-        buf[ret] = '\0';
-      } else {
-        // TODO: There is probably more data that needs to be read
-        buf[WU_BUFSIZE - 1] = '\0';
-      }
-#ifdef ENERGYMON_WATTSUP_DEBUG
-      fprintf(stdout, "Read %d characters:\n%s\n", ret, buf);
-#endif
-      // sometimes we get multiple (or at least parts of old) packets
-      // skip over old data to beginning of most recent packet
-      tok = strrchr(buf, '#');
-      if (tok) {
-        // TODO: It's possible that this is not a complete packet
-        tok = strtok_r(tok, ",", &saveptr);
-        for (i = 0; tok; i++) {
-          // 1st index = command, starting with a '#'
-          // 2nd index = subcommand (should be a '-', but we don't care)
-          if (i == WU_POWER_INDEX) {
-            // treat this is as the average power since last successful read
-            errno = 0;
-            tmp = strtoul(tok, NULL, 0);
-            if (errno) {
-              // keep old value of deciwatts
-              perror("wattsup_poll_sensors:strtoul");
-            } else {
-              state->deciwatts = tmp;
-            }
-            break;
-          }
-          tok = strtok_r(NULL, ",", &saveptr);
-        }
-        if (i != WU_POWER_INDEX) {
-          fprintf(stderr, "Failed to get power from WattsUp data: %s\n", buf);
-        }
-      } else {
-        fprintf(stderr, "Bad data from WattsUp: %s\n", buf);
-      }
+    if ((pstart = data_packet_read(state->ctx, buf, sizeof(buf), &state->poll))) {
+      data_packet_parse(pstart, &state->deciwatts);
     }
-
-    // WattsUps are cranky - don't read for a whole second
     if (state->use_estimates) {
       lock_acquire(&state->lock);
     }
@@ -153,10 +221,7 @@ static void* wattsup_poll_sensors(void* args) {
     if (state->use_estimates) {
       lock_release(&state->lock);
     }
-#ifndef __ANDROID__
-    pthread_setcancelstate(PTHREAD_CANCEL_ENABLE, &dummy_old_state);
-#endif
-    energymon_sleep_us(WU_MIN_INTERVAL_US, &state->poll);
+    wattsup_thread_sleep_us(WU_POLL_INTERVAL_US, &state->poll);
   }
   return (void*) NULL;
 }
@@ -168,7 +233,7 @@ static int wattsup_flush_read(energymon_wattsup_ctx* ctx) {
   int i;
   int ret;
   // try to get one good data packet from the device
-  for (i = 0; i <= WU_MAX_RETRIES; i++) {
+  for (i = 0; i <= WU_INIT_MAX_RETRIES; i++) {
     if ((ret = wattsup_read(ctx, buf, WU_BUFSIZE)) < 0) {
       return -1;
     }
@@ -184,7 +249,7 @@ static int wattsup_flush_read(energymon_wattsup_ctx* ctx) {
     if (strrchr(buf, '#') && strrchr(buf, ';')) {
       break;
     }
-    if (i == WU_MAX_RETRIES) {
+    if (i == WU_INIT_MAX_RETRIES) {
       return -1; // avoid unnecessary sleep
     }
     energymon_sleep_us(WU_MIN_INTERVAL_US, &IGNORE_INTERRUPT);
@@ -251,9 +316,6 @@ int energymon_init_wattsup(energymon* em) {
     errno = err_save;
     return -1;
   }
-
-  // give the polling thread a head start
-  energymon_sleep_us(WU_MIN_INTERVAL_US / 100, &state->poll);
 
   em->state = state;
   return 0;