+++ zcu102/energymon-zcu102.c
@@ -210,9 +210,9 @@ static void* zcu102_poll_sensors(void* args) {
 #endif
       state->total_uj += delta_uj;
     }
-    // sleep for the update interval of the sensors (minus most overhead)
-    if (state->read_delay_us > exec_us) {
-      energymon_sleep_us(state->read_delay_us - exec_us, &state->poll_sensors);
+    // sleep for the update interval of the sensors
+    if (state->poll_sensors) {
+      energymon_sleep_us(state->read_delay_us, &state->poll_sensors);
     }
     errno = 0;
   }