+++ RELEASES.md
@@ -14,7 +14,7 @@
 ### Fixed
  * wattsup-libftdi, wattsup-libusb: Fix empty reads by always setting baud rate and serial line properties
  * wattsup-dev: Use pselect to avoid timer expired errors
- * odroid, odroid-ioctl, osp-polling: Minor correction to polling delay interval
+ * odroid, odroid-ioctl, osp-polling, zcu102: Minor correction to polling delay interval
 
 
 ## [v0.2.3] - 2017-11-19